made a board solution checker
