+++ src/test/java/touro/peg/TriangleBoardTest.java
@@ -139,80 +139,4 @@ class TriangleBoardTest
         //Then
         assertFalse(board.isBestWin());
     }
-
-    @Test
-    public void solutions()
-    {
-        // given
-        TriangleBoard triangleBoard = new TriangleBoard(0);
-
-        // when
-        ArrayList<String> solutions = new ArrayList<>();
-        triangleBoard.getSolutions().forEach(board -> solutions.add(board.toString()));
-
-        // then
-        String expectedA = """
-                    0
-                   0 0
-                  0 0 0
-                 0 0 0 0
-                0 0 1 0 0
-                """;
-
-        String expectedB = """
-                    0
-                   0 0
-                  0 0 0
-                 0 0 0 1
-                0 0 0 0 0
-                """;
-
-        String expectedC = """
-                    1
-                   0 0
-                  0 0 0
-                 0 0 0 0
-                0 0 0 0 0
-                """;
-        assertTrue(solutions.contains(expectedA));
-        assertTrue(solutions.contains(expectedB));
-        assertTrue(solutions.contains(expectedC));
-
-    }
-
-    @Test
-    public void bestSolutions()
-    {
-        // given
-        TriangleBoard triangleBoard = new TriangleBoard(0);
-
-        // when
-        TriangleBoard bestSolutions = triangleBoard.getBestSolutions().get(0);
-
-        // then
-        String expected = """
-                    1
-                   0 0
-                  0 0 0
-                 0 0 0 0
-                0 0 0 0 0
-                """;
-        assertEquals(expected, bestSolutions.toString());
-    }
-
-    @Test
-    public void pathToBestSolutions()
-    {
-        // given
-        TriangleBoard triangleBoard = new TriangleBoard(0);
-
-        // when
-        List<TriangleBoardTree.TriangleTreeNode> firstPath
-                = triangleBoard.getPathsToBestSolutions().get(0);
-        int numMovesToReachFirst = 14;
-
-        // then
-        //System.out.println(firstPath);
-        assertEquals(numMovesToReachFirst, firstPath.size()); // expected, actual
-    }
 }
